+++ ssv/runner_state.go
@@ -18,7 +18,7 @@ type State struct {
 	PostConsensusContainer *PartialSigContainer
 	RunningInstance        *qbft.Instance
 	// Quorum is the number of min signatures needed for quorum
-	Quorum       int
+	Quorum       uint64
 	DecidedValue []byte
 	// CurrentDuty is the duty the node pulled locally from the beacon node, might be different from decided duty
 	StartingDuty types.Duty
@@ -26,7 +26,7 @@ type State struct {
 	Finished bool // Finished marked true when there is a full successful cycle (pre, consensus and post) with quorum
 }
 
-func NewRunnerState(quorum int, duty types.Duty) *State {
+func NewRunnerState(quorum uint64, duty types.Duty) *State {
 	return &State{
 		PreConsensusContainer:  NewPartialSigContainer(quorum),
 		PostConsensusContainer: NewPartialSigContainer(quorum),
